extract SerializableConfiguration to util
